Refactor content library components and audio player
- Removed ContentLibraryDrawer component and replaced it with modular components: ContentListItem, ContentSearchBar, NowPlayingBanner, and NowPlayingFooter for better separation of concerns.
- Introduced TransportControls and SpeedSelector components for audio playback controls, enhancing reusability and maintainability.
- Added utility functions for time formatting in a new formatTime.js file.
- Updated audio player components to utilize new transport controls and speed selector, improving user experience and code organization.
